Move `tools` module to `deprecated` namespace
