--- docs/api/tools.md
@@ -1,1 +1,0 @@
-::: pydantic.tools
