+++ pydantic/__init__.py
@@ -13,11 +13,11 @@ from .analyzed_type import AnalyzedType
 from .config import ConfigDict, Extra
 from .decorators import field_serializer, field_validator, model_serializer, root_validator, validator
 from .deprecated.config import BaseConfig
+from .deprecated.tools import *
 from .errors import *
 from .fields import Field, PrivateAttr
 from .main import *
 from .networks import *
-from .tools import *
 from .types import *
 from .validate_call import validate_call
 from .version import VERSION
